Move component tests to component folder
