+++ application/ui/playwright.config.ts
@@ -9,8 +9,7 @@ const CI = !!process.env.CI;
  * See https://playwright.dev/docs/test-configuration.
  */
 export default defineConfig({
-    testDir: './tests',
-    testIgnore: '**/e2e/**',
+    testDir: './tests/component',
     /* Run tests in files in parallel */
     fullyParallel: true,
     /* Fail the build on CI if you accidentally left test.only in the source code. */
