+++ application/ui/tests/component/inference.spec.ts
