+++ application/ui/tests/component/project.spec.ts
@@ -4,7 +4,7 @@
 import { Page } from '@playwright/test';
 import { HttpResponse } from 'msw';
 
-import { getMockedProject } from '../mocks/mock-project';
+import { getMockedProject } from '../../mocks/mock-project';
 import { expect, http, test } from './fixtures';
 
 const fillProjectForm = async ({
